wip: component arranging
